# combfullui/ui/swagger.yaml
openapi: 3.0.0
info:
  title: Public Comb Protocol API
  description: "
  # Public Comb Protocol #
  
  It is a protocol for a Haircomb Core to communicate with its clients. Haircomb Core is the software which maintains the haircomb commitment set for the haircomb crypto currency.
  
  # Authentication #
  
  Authentication is not in use.

  # Api results content-type #
  
  Actual contents of the response may be JSONP encoded. Example in this docs is
  
  ```
  {\"Success\":true,...}
  ```
  
  Actual value received from the software may be
  
  ```
  C({\"Success\":true,...});
  ```
  
  The function name C in this case might be arbitrary uppercase and lowercase ascii letters. Swagger documentation, however, does not format jsonp nicely, that's why I use application/json for reference.

  # Api results error handling #

  In case of errors Success false will be reported. The HTTP code will be 203 in case of error, and 200 in the general case, allowing even clients unaware about http error codes to detect the error.

  # Servers #
  
  ## Cloud instance (mainnet). ##
  
  - [https://core.haircomb.org](https://core.haircomb.org)
    - Main production server
    
  - [http://core.haircomb.org:21212](http://core.haircomb.org:21212)
    - Main production server
    - No Nginx and No HTTPS overhead

  - [http://72.5.43.45:21212](http://72.5.43.45:21212)
    - Main production server
    - No DNS, No Nginx and No HTTPS overhead
    
...
    - Main production server
...
  
...

...

...

...
  
...

...

  "
...
servers:
...
paths:
...
  /*/main.{json}:
    get:
      summary: Main Page Controller
      description: Loads the main page information. This contains information regarding the node syncing status, as well as network to which the node is synchronized.
      operationId: v0.MainPageController
      tags:
        - Backend (1 dot)
      parameters:
        - in: path
          name: json
          schema: {}
          required: true
          description: Arbitrary string not containing dots. It is ignored by the server, serves as an extension for the jsonp resoruce. Usual values are js, json, jsonp.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MainPageInfo'
              example:
                Testnet: false
                OutOfSync: false
                LastBtcHeight: 811035
                CryptoFingerPrint: dfc83b86041348a256f8cf3f24b1888e8c5ed363ad546299f377015eb094d908
                CommitmentsCount: 21129936
                Accelerator: 999AB5E53D0352224F1C656D756710EF5A2DFE72B48A3DD9EA7E04D8BC5C449A
                BlockHash: '000000000000000000003441ad183b60b2280b0cf5ecb7566d6e6174b9e48551'
                SumExistence: 44179636923846
                SumRemaining: 1103730671845172
                BlockHeight: 811035
                Success: true
                P2WSHCount: 31766113
                ShutdownButton: false
        '504':
          description: Nginx Gateway Timeout (Service is Starting)
          content:
            text/html:
              schema:
                $ref: '#/components/schemas/NginxGatewayTimeout'
        '203':
          description: Processing Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingError'
              example:
                {"Success":false}
...
  /*/combfullui.{wasm}:
    get:
      summary: Static Wasm Resource Loader
      operationId: v0.StaticWasmResourceLoader
      tags:
        - Frontend (1 dot)
      parameters:
        - in: path
          name: wasm
          schema:
            type: string
          required: true
          description: Arbitrary string not containing dots. It is ignored by the server, serves as an extension for the wasm resoruce. Usual value is wasm.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
...
components:
  schemas:
    MainPageInfo:
      items:
        $ref: '#/components/schemas/MainPageInfo'
      type: object
      title: MainPageInfo
      properties:
        Testnet:
          type: boolean
        OutOfSync:
          type: boolean
        LastBtcHeight:
          type: integer
        CryptoFingerPrint:
          type: string
        CommitmentsCount:
          type: integer
        Accelerator:
          type: string
        BlockHash:
          type: string
        SumExistence:
          type: integer
        SumRemaining:
          type: integer
        BlockHeight:
          type: integer
        Success:
          type: boolean
        P2WSHCount:
          type: integer
        ShutdownButton:
          type: boolean
      required:
        - Testnet
        - OutOfSync
        - LastBtcHeight
        - CryptoFingerPrint
        - CommitmentsCount
        - Accelerator
        - BlockHash
        - SumExistence
        - SumRemaining
        - BlockHeight
        - Success
        - P2WSHCount
        - ShutdownButton
...
    NginxGatewayTimeout:
      items:
        $ref: '#/components/schemas/NginxGatewayTimeout'
      type: object
      title: NginxGatewayTimeout
    ProcessingError:
      properties:
        Success:
          type: boolean
          title: Boolean
          example: false
      items:
        $ref: '#/components/schemas/ProcessingError'
      type: object
      title: ProcessingError